[MISC] move alphabet_base out of detail
