+++ include/seqan3/alphabet/aminoacid/aminoacid_base.hpp
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 #include <seqan3/alphabet/detail/convert.hpp>
 #include <seqan3/alphabet/aminoacid/concept.hpp>
 #include <seqan3/io/stream/char_operations.hpp>
