+++ include/seqan3/alphabet/cigar/cigar_op.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 
 // ------------------------------------------------------------------
 // cigar_op
