+++ include/seqan3/alphabet/composite/semialphabet_any.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 
 namespace seqan3
 {
