+++ include/seqan3/alphabet/gap/gap.hpp
@@ -15,7 +15,7 @@
 
 #include <cassert>
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 
 namespace seqan3
 {
