+++ include/seqan3/alphabet/nucleotide/nucleotide_base.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 #include <seqan3/alphabet/detail/convert.hpp>
 #include <seqan3/alphabet/nucleotide/concept.hpp>
 #include <seqan3/io/stream/char_operations.hpp>
