+++ include/seqan3/alphabet/structure/dot_bracket3.hpp
@@ -14,7 +14,7 @@
 
 #include <vector>
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 #include <seqan3/alphabet/structure/concept.hpp>
 #include <seqan3/io/stream/char_operations.hpp>
