+++ include/seqan3/alphabet/structure/wuss.hpp
@@ -15,7 +15,7 @@
 #include <cmath>
 #include <vector>
 
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 #include <seqan3/alphabet/structure/concept.hpp>
 #include <seqan3/io/stream/char_operations.hpp>
