+++ test/snippet/alphabet/detail/alphabet_base.cpp
@@ -1,5 +1,5 @@
 #include <seqan3/alphabet/concept.hpp>
-#include <seqan3/alphabet/detail/alphabet_base.hpp>
+#include <seqan3/alphabet/alphabet_base.hpp>
 
 using namespace seqan3;
